requirements.txt
fastapi
uvicorn
python-docx
python-dotenv
python-multipart
sarvamai

# Test dependencies
pytest
httpx
